test(learning): use f32 candidate confidence

+++ crates/davinci-coding-agent/src/native_extensions/learning/store.rs
@@ -504,7 +504,7 @@ mod tests {
         let mut store = LearningStore::open(dir.path().to_path_buf()).unwrap();
         let mut candidate = fixture_candidate("cand-bloated");
         for i in 0..100 {
-            candidate.confidence = (i as f64) / 100.0;
+            candidate.confidence = (i as f32) / 100.0;
             store.upsert_candidate(candidate.clone()).unwrap();
         }
         drop(store);
